Fix import cycle (#27)
* Fix import cycle causing the build to fail

* Add membrane.ts
